Fix registry key deletion when compiled in 32bits mode
When running in 32 bits mode, windows present a slightly different
registry to an application [1]. This means that deleting a key from there
doesn't delete it properly. The registry crate doesn't have a way to
pass `Wow6464Key` to its delete calls so we have to open a key to delete
it which is slightly inconvenient but works.

[1]: https://docs.microsoft.com/en-us/windows/win32/api/winreg/nf-winreg-regdeletekeyw